+++ src/main/java/de/fh_zwickau/informatik/sensor/ZWayConstants.java
@@ -26,6 +26,7 @@ public class ZWayConstants {
     public static final String PATH_PROFILES = "profiles";
     public static final String PATH_NOTIFICATIONS = "notifications";
     public static final String PATH_DEVICE_HISTORY = "history";
+    public static final String PATH_ICONS = "icons";
 
     public static final String REMOTE_PATH_LOGIN = "zboxweb";